feat: extend short flag counts to int64, fix single case
The '-vvv' example where that returns '3' for an int flag with short
'v', only worked for Int, not Int64.

While here we also fix the case '-v -vvv'. Here we want the value '3'
again, since last cluster wins, but this would fail as it'd try to
interpret '-vvv' as an int arg to the '-v' flag preceding it.

'-vv -vvv' would work correctly, because it'd not expect the next think
after '-vv' to be a flag, but this should (arguably) also apply to a
single '-v'. It's a bit debateable because it arguably leaves too much
room for error.

A consequence of this change is that "myscript -v" is valid and returns
'1' for 'v'. We might need to change this in the future, if we find
users find to too easy to make errors with.